demos/03-crewai-avp/requirements.txt
avp-sdk>=0.1.0
crewai>=0.28.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
anthropic>=0.18.0
openai>=1.0.0
